+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/treeview/FolderTreeItemEntry.java
@@ -1,6 +1,8 @@
 package com.kaylerrenslow.armaDialogCreator.gui.main.treeview;
 
 import com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.treeView.CellType;
+import javafx.scene.control.TreeItem;
+import javafx.scene.control.TreeView;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -10,5 +12,13 @@ public class FolderTreeItemEntry extends TreeItemEntry {
 
 	public FolderTreeItemEntry(@NotNull String folderText) {
 		super(folderText, CellType.FOLDER, EditorComponentTreeView.createFolderIcon());
+	}
+
+	@Override
+	@SuppressWarnings("unchecked")
+	public void duplicate(@NotNull TreeView<? extends TreeItemEntry> treeView) {
+		treeView.getRoot().getChildren().add(new TreeItem(
+				new FolderTreeItemEntry(this.getText() + "copy")
+		));
 	}
 }